LayoutTests/fast/frames/seamless/seamless-css-cascade-expected.txt

PASS window.getComputedStyle(one).color is "rgb(255, 0, 0)"
PASS window.getComputedStyle(two).color is "rgb(128, 0, 128)"
PASS window.getComputedStyle(three).color is "rgb(255, 255, 255)"
PASS window.getComputedStyle(rootElement).color is "rgb(255, 165, 0)"
